+++ scheduler/Scheduler.cpp
@@ -1829,49 +1829,61 @@ std::list<common::dataStructures::QueueAndMountSummary> Scheduler::getQueuesAndM
 
   for (auto & pm: mountDecisionInfo->potentialMounts) {
     // Find or create the relevant entry.
-    auto &summary = common::dataStructures::QueueAndMountSummary::getOrCreateEntry(ret, pm.type, pm.tapePool, pm.vid, vid_to_logical_library);
+    common::dataStructures::QueueAndMountSummary* summary;
+    try {
+      summary = common::dataStructures::QueueAndMountSummary::getOrCreateEntry(ret, pm.type, pm.tapePool, pm.vid,
+        vid_to_logical_library);
+    } catch (exception::Exception & ex) {
+      log::ScopedParamContainer params(lc);
+      params.add("vid", pm.vid);
+      params.add("tapepool", pm.tapePool);
+      params.add("errorMessage", ex.getMessage().str());
+      lc.log(log::WARNING, "In Scheduler::getQueuesAndMountSummaries(): "
+        "got an exception trying to create a queue summary entry. Invalid mount type. Skipping.");
+      continue;
+    }
     switch (pm.type) {
     case common::dataStructures::MountType::ArchiveForUser:
     case common::dataStructures::MountType::ArchiveForRepack:
-      summary.mountPolicy.archivePriority = pm.priority;
-      summary.mountPolicy.archiveMinRequestAge = pm.minRequestAge;
-      summary.bytesQueued = pm.bytesQueued;
-      summary.filesQueued = pm.filesQueued;
-      summary.oldestJobAge = time(nullptr) - pm.oldestJobStartTime ;
-      summary.youngestJobAge = time(nullptr) - pm.youngestJobStartTime;
+      summary->mountPolicy.archivePriority = pm.priority;
+      summary->mountPolicy.archiveMinRequestAge = pm.minRequestAge;
+      summary->bytesQueued = pm.bytesQueued;
+      summary->filesQueued = pm.filesQueued;
+      summary->oldestJobAge = time(nullptr) - pm.oldestJobStartTime ;
+      summary->youngestJobAge = time(nullptr) - pm.youngestJobStartTime;
       if (pm.mountPolicyNames) {
-        summary.mountPolicies = pm.mountPolicyNames.value();
+        summary->mountPolicies = pm.mountPolicyNames.value();
       }
       if (pm.highestPriorityMountPolicyName) {
-        summary.highestPriorityMountPolicy = pm.highestPriorityMountPolicyName.value();
+        summary->highestPriorityMountPolicy = pm.highestPriorityMountPolicyName.value();
       }
       if (pm.lowestRequestAgeMountPolicyName) {
-        summary.lowestRequestAgeMountPolicy = pm.lowestRequestAgeMountPolicyName.value();
+        summary->lowestRequestAgeMountPolicy = pm.lowestRequestAgeMountPolicyName.value();
       }
       break;
     case common::dataStructures::MountType::Retrieve:
       // TODO: we should remove the retrieveMinRequestAge if it's redundant, or rename pm.minArchiveRequestAge.
-      summary.mountPolicy.retrieveMinRequestAge = pm.minRequestAge;
-      summary.mountPolicy.retrievePriority = pm.priority;
-      summary.bytesQueued = pm.bytesQueued;
-      summary.filesQueued = pm.filesQueued;
-      summary.oldestJobAge = time(nullptr) - pm.oldestJobStartTime ;
-      summary.youngestJobAge = time(nullptr) - pm.youngestJobStartTime;
+      summary->mountPolicy.retrieveMinRequestAge = pm.minRequestAge;
+      summary->mountPolicy.retrievePriority = pm.priority;
+      summary->bytesQueued = pm.bytesQueued;
+      summary->filesQueued = pm.filesQueued;
+      summary->oldestJobAge = time(nullptr) - pm.oldestJobStartTime ;
+      summary->youngestJobAge = time(nullptr) - pm.youngestJobStartTime;
       if (pm.mountPolicyNames) {
-        summary.mountPolicies = pm.mountPolicyNames.value();
+        summary->mountPolicies = pm.mountPolicyNames.value();
       }
       if (pm.highestPriorityMountPolicyName) {
-        summary.highestPriorityMountPolicy = pm.highestPriorityMountPolicyName.value();
+        summary->highestPriorityMountPolicy = pm.highestPriorityMountPolicyName.value();
       }
       if (pm.lowestRequestAgeMountPolicyName) {
-        summary.lowestRequestAgeMountPolicy = pm.lowestRequestAgeMountPolicyName.value();
+        summary->lowestRequestAgeMountPolicy = pm.lowestRequestAgeMountPolicyName.value();
       }
       if (pm.sleepingMount) {
         common::dataStructures::QueueAndMountSummary::SleepForSpaceInfo sfsi;
         sfsi.startTime = pm.sleepStartTime;
         sfsi.diskSystemName = pm.diskSystemSleptFor;
         sfsi.sleepTime = pm.sleepTime;
-        summary.sleepForSpaceInfo = sfsi;
+        summary->sleepForSpaceInfo = sfsi;
       }
       break;
     default:
@@ -1879,18 +1891,29 @@ std::list<common::dataStructures::QueueAndMountSummary> Scheduler::getQueuesAndM
     }
   }
   for (auto & em: mountDecisionInfo->existingOrNextMounts) {
-    auto &summary = common::dataStructures::QueueAndMountSummary::getOrCreateEntry(ret, em.type, em.tapePool, em.vid, vid_to_logical_library);
+    common::dataStructures::QueueAndMountSummary* summary;
+    try {
+      summary = common::dataStructures::QueueAndMountSummary::getOrCreateEntry(ret, em.type, em.tapePool, em.vid,
+        vid_to_logical_library);
+    } catch (exception::Exception & ex) {
+      log::ScopedParamContainer params(lc);
+      params.add("driveName", em.driveName);
+      params.add("errorMessage", ex.getMessage().str());
+      lc.log(log::WARNING, "In Scheduler::getQueuesAndMountSummaries(): "
+        "got an exception trying to create a queue summary entry. Invalid mount type. Skipping.");
+      continue;
+    }
     switch (em.type) {
     case common::dataStructures::MountType::ArchiveForUser:
     case common::dataStructures::MountType::ArchiveForRepack:
     case common::dataStructures::MountType::Retrieve:
       if (em.currentMount)
-        summary.currentMounts++;
+        summary->currentMounts++;
       /*else
-        summary.nextMounts++;*/
-      summary.currentBytes += em.bytesTransferred;
-      summary.currentFiles += em.filesTransferred;
-      summary.averageBandwidth = em.averageBandwidth;
+        summary->nextMounts++;*/
+      summary->currentBytes += em.bytesTransferred;
+      summary->currentFiles += em.filesTransferred;
+      summary->averageBandwidth = em.averageBandwidth;
       break;
     default:
       break;